Изменилась схема именования сгенерированных файлов (#77)
Для всех файлов исходных текстов создаётся пара файлов filename.rasl
и filename.rasl.cpp, для файлов с нативными функциями — filename.cpp.

При этом из двух файлов (*.rasl и *.rasl.cpp) первичным является файл
с расширением .rasl, поскольку в дальнейшем планируется отказаться от второго
файла.

Старая схема, соответственно, была: filename.rasl+filename.cpp ++
++ filename.native.cpp.

В соответствие с новой схемой обновлены скрипты атотестов и скрипты src/make.*,
(они теперь перед выполнением компиляции очищают соответствующую подпапку
в build/*).

Файлы refalrts.rasl и refalrts.rasl.cpp созданы на всякий случай для srmake.

+++ src/srlib/make.sh
@@ -2,18 +2,17 @@
 
 prepare_srclib() {
   ../../bin/srefc-core $SREFC_FLAGS $1
-  grep '//FROM' < $1.sref >> $1.cpp
-  grep '//FROM' < $1.sref >> $1.cpp.froms
-  [ -e $1.native.cpp ] && mv $1.native.cpp ../../srlib
+  grep '//FROM' < $1.sref >> $1.rasl.froms
+  [ -e $1.cpp ] && mv $1.cpp ../../srlib
   mv $1.rasl ../../srlib
-  mv $1.cpp ../../srlib
-  mv $1.cpp.froms ../../srlib
+  mv $1.rasl.cpp ../../srlib
+  mv $1.rasl.froms ../../srlib
   cp $1.sref ../../srlib/src
 }
 
 
 mkdir -p ../../srlib/src
-cp LICENSE *.h *.cpp ../../srlib
+cp LICENSE *.h *.rasl *.cpp ../../srlib
 prepare_srclib Library
 prepare_srclib LibraryEx
 prepare_srclib GetOpt
